LayoutTests/media/media-play-promise-reject-load-abort-expected.txt


RUN(mediaElement.play().then(failTest).catch(promiseRejected))
RUN(mediaElement.src = findMediaFile("video", "content/test"))
Promise rejected. OK
EXPECTED (error.name == 'AbortError') OK
END OF TEST
